fix: get attestations with filtered schema in same query as that of hypercert
