+++ app/hypercert/[hypercertId]/components/Attestations/index.tsx
@@ -49,7 +49,7 @@ const Attestations = () => {
 					) : (
 						validAttestations.slice(0, 2).map((attestation) => {
 							return (
-								<Attestation key={attestation.id} attestation={attestation} />
+								<Attestation key={attestation.uid} attestation={attestation} />
 							);
 						})
 					)}
